Fix find() results from not being JSON.parsed
Schema~find() was returning an Array of JSON strings,
not objects.
This was because JSON.parse was not being called correctly:

```js
for(var i = 0; i < resp.length; i++) {
  resp[i] = JSON.parse(resp[i]); // resp is still a string?
}
return resp;
```

However, when I changed the code to use to Array~map() it worked.

```js
return resp.map(JSON.parse);
```